+++ src/error.ts
@@ -1,4 +1,16 @@
+export enum ErrorMessages {
+  InvalidPassword = 'Invalid email or password',
+  UserAlreadyExists = 'User already exists',
+  CodeExpired = 'Code expired',
+  InvalidCode = 'Invalid code',
+  InvalidEmail = 'Invalid email',
+  MissingPassword = 'Missing password',
+  MissingEmail = 'Missing email',
+  MissingCode = 'Missing code',
+}
+
 export class AuthError extends Error {
+  static Messages = ErrorMessages;
   constructor(message: string, public statusCode: number) {
     super(message);
   }